V3

+++ openmusic-v3/src/services/postgres/PlaylistsService.js
@@ -5,9 +5,10 @@ const NotFoundError = require('../../exceptions/NotFoundError');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class PlaylistsService {
-  constructor(collaborationsService) {
+  constructor(collaborationsService, cacheService) {
     this.pool = new Pool();
     this.collaborationsService = collaborationsService;
+    this.cacheService = cacheService;
   }
 
   async verifyPlaylistOwner(playlistId, userId) {
@@ -75,20 +76,28 @@ class PlaylistsService {
       throw new InvariantError('Failed to add playlist.');
     }
 
+    await this.cacheService.delete(`playlists:${owner}`);
+
     return result.rows[0].id;
   }
 
   async getPlaylistsByUserId(userId) {
-    const query = {
-      text: `SELECT DISTINCT playlists.id, playlists.name, users.username FROM playlists
-      LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
-      JOIN users ON playlists.owner = users.id
-      WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
-      values: [userId],
-    };
+    try {
+      const result = await this.cacheService.get(`playlists:${userId}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT DISTINCT playlists.id, playlists.name, users.username FROM playlists
+        LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
+        JOIN users ON playlists.owner = users.id
+        WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
+        values: [userId],
+      };
 
-    const result = await this.pool.query(query);
-    return result.rows;
+      const result = await this.pool.query(query);
+      await this.cacheService.set(`playlist:${userId}`, JSON.stringify(result.rows));
+      return result.rows;
+    }
   }
 
   async deleteAllSongsFromPlaylist(playlistId) {
@@ -98,13 +107,14 @@ class PlaylistsService {
     };
 
     await this.pool.query(query);
+    await this.cacheService.delete(`playlistsongs:${playlistId}`);
   }
 
   async deletePlaylistById(playlistId) {
     await this.deleteAllSongsFromPlaylist(playlistId);
 
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [playlistId],
     };
 
@@ -113,6 +123,8 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new NotFoundError('Cannot delete playlist. Playlist id not found.');
     }
+
+    await this.cacheService.delete(`playlists:${result.rows[0].owner}`);
   }
 
   async addSongToPlaylist({
@@ -130,18 +142,27 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new InvariantError('Failed to add song to playlist.');
     }
+
+    await this.cacheService.delete(`playlistsongs:${playlistId}`);
   }
 
   async getSongsByPlaylistId(playlistId) {
-    const query = {
-      text: `SELECT songs.id, songs.title, songs.performer FROM playlistsongs
-      JOIN songs on playlistsongs.song_id = songs.id
-      WHERE playlistsongs.playlist_id = $1`,
-      values: [playlistId],
-    };
+    try {
+      const result = await this.cacheService.get(`playlistsongs:${playlistId}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT songs.id, songs.title, songs.performer FROM playlistsongs
+        JOIN songs on playlistsongs.song_id = songs.id
+        WHERE playlistsongs.playlist_id = $1`,
+        values: [playlistId],
+      };
 
-    const result = await this.pool.query(query);
-    return result.rows;
+      const result = await this.pool.query(query);
+      await this.cacheService.set(`playlistsongs:${playlistId}`, JSON.stringify(result.rows));
+
+      return result.rows;
+    }
   }
 
   async deleteSongFromPlaylist({
@@ -158,6 +179,8 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new InvariantError('Failed to delete song from playlist.');
     }
+
+    await this.cacheService.delete(`playlistsongs:${playlistId}`);
   }
 }
